rename API server to XKCD
