+++ internal-xkcd/core/domain/weights.go
