+++ internal-xkcd/core/domain/words.go
